Refactor certificate loading methods into a separate CertificateLoader class
