+++ src/EventStore.Core/VNodeBuilder.cs
@@ -666,33 +666,7 @@ namespace EventStore.Core {
 			string certificatePath,
 			string privateKeyPath,
 			string password) {
-
-			if (string.IsNullOrEmpty(privateKeyPath)) {
-				try {
-					_certificate = new X509Certificate2(certificatePath, password);
-				} catch (CryptographicException exc) {
-					throw new AggregateException("Error loading certificate file. Please verify that the correct password has been provided via the `CertificatePassword` option.", exc);
-				}
-
-				if (!_certificate.HasPrivateKey) {
-					throw new Exception("Expect certificate to contain a private key. " +
-					                    "Please either provide a certificate that contains one or set the private key" +
-					                    " via the `CertificatePrivateKeyFile` option.");
-				}
-				return this;
-			}
-			
-			var privateKey = Convert.FromBase64String(string.Join(string.Empty, File.ReadAllLines(privateKeyPath)
-				.Skip(1)
-				.SkipLast(1)));
-			
-			using var rsa = RSA.Create();
-			rsa.ImportRSAPrivateKey(new ReadOnlySpan<byte>(privateKey), out _);
-			
-			using var publicCertificate = new X509Certificate2(certificatePath, password);
-			using var publicWithPrivate = publicCertificate.CopyWithPrivateKey(rsa);
-			
-			_certificate = new X509Certificate2(publicWithPrivate.Export(X509ContentType.Pfx));
+			_certificate = CertificateLoader.FromFile(certificatePath, privateKeyPath, password);
 			return this;
 		}
 
@@ -704,28 +678,7 @@ namespace EventStore.Core {
 		public VNodeBuilder WithTrustedRootCertificatesPath(
 			string trustedRootCertificatesPath) {
 			Ensure.NotNullOrEmpty(trustedRootCertificatesPath, "trustedRootCertificatesPath");
-
-			var certCollection = new X509Certificate2Collection();
-			var files = Directory.GetFiles(trustedRootCertificatesPath);
-			var acceptedExtensions = new[] { ".crt", ".cert", ".cer", ".pem", ".der" };
-			foreach (var file in files) {
-				var fileInfo = new FileInfo(file);
-				var extension = fileInfo.Extension;
-				if (acceptedExtensions.Contains(extension)) {
-					try {
-						var cert = new X509Certificate2(File.ReadAllBytes(file));
-						certCollection.Add(cert);
-						_log.Information("Trusted root certificate file loaded: {file}", fileInfo.Name);
-					} catch (Exception exc) {
-						throw new AggregateException($"Error loading trusted root certificate file: {file}", exc);
-					}
-				}
-			}
-
-			if(certCollection.Count == 0)
-				throw new Exception($"No trusted root certificates were loaded from: {trustedRootCertificatesPath}");
-
-			_trustedRootCerts = certCollection;
+			_trustedRootCerts = CertificateLoader.LoadCertificateCollection(trustedRootCertificatesPath);
 			return this;
 		}
 
@@ -1182,8 +1135,7 @@ namespace EventStore.Core {
 		/// <returns>A <see cref="VNodeBuilder"/> with the options set</returns>
 		public VNodeBuilder WithServerCertificateFromStore(StoreLocation storeLocation, StoreName storeName,
 			string certificateSubjectName, string certificateThumbprint) {
-			var store = new X509Store(storeName, storeLocation);
-			_certificate = GetCertificateFromStore(store, certificateSubjectName, certificateThumbprint);
+			_certificate = CertificateLoader.FromStore(storeLocation, storeName, certificateSubjectName, certificateThumbprint);
 			return this;
 		}
 
@@ -1196,54 +1148,9 @@ namespace EventStore.Core {
 		/// <returns>A <see cref="VNodeBuilder"/> with the options set</returns>
 		public VNodeBuilder WithServerCertificateFromStore(StoreName storeName, string certificateSubjectName,
 			string certificateThumbprint) {
-			var store = new X509Store(storeName);
-			_certificate = GetCertificateFromStore(store, certificateSubjectName, certificateThumbprint);
+			_certificate = CertificateLoader.FromStore(storeName, certificateSubjectName, certificateThumbprint);
 			return this;
 		}
-
-		private X509Certificate2 GetCertificateFromStore(X509Store store, string certificateSubjectName,
-			string certificateThumbprint) {
-			try {
-				store.Open(OpenFlags.OpenExistingOnly);
-			} catch (Exception exc) {
-				throw new Exception(
-					string.Format("Could not open certificate store '{0}' in location {1}'.", store.Name,
-						store.Location), exc);
-			}
-
-			if (!string.IsNullOrWhiteSpace(certificateThumbprint)) {
-				var certificates = store.Certificates.Find(X509FindType.FindByThumbprint, certificateThumbprint, true);
-				if (certificates.Count == 0)
-					throw new Exception(string.Format("Could not find valid certificate with thumbprint '{0}'.",
-						certificateThumbprint));
-
-				//Can this even happen?
-				if (certificates.Count > 1)
-					throw new Exception(string.Format("Could not determine a unique certificate from thumbprint '{0}'.",
-						certificateThumbprint));
-
-				return certificates[0];
-			}
-
-			if (!string.IsNullOrWhiteSpace(certificateSubjectName)) {
-				var certificates =
-					store.Certificates.Find(X509FindType.FindBySubjectName, certificateSubjectName, true);
-				if (certificates.Count == 0)
-					throw new Exception(string.Format("Could not find valid certificate with subject name '{0}'.",
-						certificateSubjectName));
-
-				//Can this even happen?
-				if (certificates.Count > 1)
-					throw new Exception(string.Format(
-						"Could not determine a unique certificate from subject name '{0}'.", certificateSubjectName));
-
-				return certificates[0];
-			}
-
-			throw new ArgumentException(
-				"No thumbprint or subject name was specified for a certificate, but a certificate store was specified.");
-		}
-
 
 		/// <summary>
 		/// Sets the transaction file chunk size. Default is <see cref="TFConsts.ChunkSize"/>
